migrate to postgresql
this took a lot of work. i thought i could optimize sqlite by doing everything in transaction and only committing from time to time, however because of sqlite locking (i thought at first it would support multithreading fine...) it resulted in a bunch of database is locked/infinite waits. i tried some other stuff during a few days and resorted to just use a different database... which was the solution. so welcome postgresql